uyPro/uyPro/requirements.txt
beautifulsoup4==4.11.2
itemadapter==0.7.0
newspaper3k==0.2.8
pandas==1.5.3
python_dateutil==2.8.2
pytz==2022.7.1
redis==4.6.0
Scrapy==2.11.1
scrapy_selenium==0.0.7
translators==5.9.9
Twisted==22.10.0
watchdog==2.3.1
deep-translator==1.11.4
PyExecJS==1.5.1
py3langid==0.2.2
lxml_html_clean==0.1.1
demjson3==3.0.6
jsonpath_ng==1.6.1
cloudscraper==1.2.71
certifi==2024.8.30
bcrypt==4.2.1
googletrans==4.0.2
google-genai==1.2.0
DrissionPage==4.0.4.23
convertdate==2.4.0
lxml==5.3.1